07 DALLE Demo/requirements.txt
openai==0.28
python-dotenv
flask